write tests for buttongroup

+++ superintendent/controls/buttongroup.py
@@ -73,7 +73,7 @@ class ButtonGroup(widgets.HBox):
     @traitlets.validate("button_width")
     def _valid_value(self, proposal: Dict):
         if isinstance(proposal["value"], Number) and proposal["value"] <= 1:
-            return "{}%".format(100 * proposal["value"])
+            return "{}%".format(int(100 * proposal["value"]))
         elif isinstance(proposal["value"], Number):
             return "{}px".format(int(proposal["value"]))
         elif isinstance(proposal["value"], str):
